+++ knf/validators/cron/validators.go
@@ -18,20 +18,21 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 var (
-	// Expression returns an error if config property contains invalid cron expression
+	// Expression returns an error if configuration property contains invalid cron
+	// expression
 	Expression = validateCronExpression
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func validateCronExpression(config knf.IConfig, prop string, value any) error {
-	confVal := config.GetS(prop)
+	v := config.GetS(prop)
 
-	if confVal == "" {
+	if v == "" {
 		return nil
 	}
 
-	_, err := cron.Parse(confVal)
+	_, err := cron.Parse(v)
 
 	if err != nil {
 		return fmt.Errorf("Property %s contains invalid cron expression: %w", prop, err)