+++ knf/validators/system/validators.go
@@ -19,32 +19,32 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 var (
-	// User returns error if config property contains name of user or UID which not
+	// User returns error if configuration property contains name of user or UID which not
 	// present on the system
 	User = validateUser
 
-	// Group returns error if config property contains name of group or GID which not
-	// present on the system
+	// Group returns error if configuration property contains name of group or GID which
+	// not present on the system
 	Group = validateGroup
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func validateUser(config knf.IConfig, prop string, value any) error {
-	userNameOrID := config.GetS(prop)
+	v := config.GetS(prop)
 
-	if userNameOrID != "" && !system.IsUserExist(userNameOrID) {
-		return fmt.Errorf("User %q is not present on the system", userNameOrID)
+	if v != "" && !system.IsUserExist(v) {
+		return fmt.Errorf("User %q is not present on the system", v)
 	}
 
 	return nil
 }
 
 func validateGroup(config knf.IConfig, prop string, value any) error {
-	groupNameOrID := config.GetS(prop)
+	v := config.GetS(prop)
 
-	if groupNameOrID != "" && !system.IsGroupExist(groupNameOrID) {
-		return fmt.Errorf("Group %q is not present on the system", groupNameOrID)
+	if v != "" && !system.IsGroupExist(v) {
+		return fmt.Errorf("Group %q is not present on the system", v)
 	}
 
 	return nil