+++ timeutil/timeutil.go
@@ -54,7 +54,7 @@ func PrettyDuration(d any) string {
 		return ""
 	}
 
-	if dur != 0 && dur < time.Second {
+	if dur != 0 && dur < SECOND {
 		return getPrettyShortDuration(dur, " ")
 	}
 
@@ -264,7 +264,7 @@ func ParseDuration(dur string, defMod ...rune) (time.Duration, error) {
 		}
 	}
 
-	return time.Second * time.Duration(result), nil
+	return SECOND * time.Duration(result), nil
 }
 
 // StartOfHour returns start of the hour
@@ -475,15 +475,20 @@ func IsWeekend(t time.Time) bool {
 }
 
 // Until returns time until given moment in given units
-func Until(t time.Time, mod time.Duration) int {
+func Until(t time.Time, unit time.Duration) int {
 	now := time.Now().In(t.Location())
-	return int(t.Sub(now) / mod)
+	return int(t.Sub(now) / unit)
 }
 
 // Since returns time since given moment in given units
-func Since(t time.Time, mod time.Duration) int {
+func Since(t time.Time, unit time.Duration) int {
 	now := time.Now().In(t.Location())
-	return int(now.Sub(t) / mod)
+	return int(now.Sub(t) / unit)
+}
+
+// DurationAs returns duration in given units
+func DurationAs(t, unit time.Duration) int {
+	return int(t / unit)
 }
 
 // FromISOWeek returns date for given week number in given year
@@ -504,25 +509,25 @@ func convertDuration(d any) (time.Duration, bool) {
 	case time.Duration:
 		return u, true
 	case int:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case int16:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case int32:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case uint:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case uint16:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case uint32:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case uint64:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case float32:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case float64:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	case int64:
-		return time.Duration(u) * time.Second, true
+		return time.Duration(u) * SECOND, true
 	}
 
 	return 0, false