+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### [13.20.2](https://kaos.sh/ek/13.20.2)
+
+- **`[knf/validators/time]`** Added timezone validator
+- **`[timeutil]`** Added method `DurationAs`
+- **`[timeutil]`** Method `Period.DurationIn` renamed to `Period.DurationAs`
+- **`[knf/validators]`** Code refactoring
+- **`[timeutil]`** Code refactoring
+
 ### [13.20.1](https://kaos.sh/ek/13.20.1)
 
 - **`[timeutil]`** Added method `Period.Stringf`